Erstellten JSON String ausgeben lassen
